Programatic Todo Additions/Todo Item design
I started to add the features so that a click on a button adds a todo
item. This will of course need to be further fleshed out so that the
user can pick what goes into the to do list. I also started to design
the actual list entries giving them a style.