Added Programatic Todo Style/Redesigned Stylesheet
Added the same style i had in XML for the todo List to the program so
that a button can activate it. Now the location is also dynamic. Also
redesigned the stylesheet so that only the stroke seperating line
appears on the bottom. This prevents a 2dp line inbetween two items,
making for a cleaner look.